Improve Vulkan and cross-API framebuffer handling
- Updated framebuffer lifecycle with cross-API abstractions for compatibility.
- Refactored `GetColorAttachmentRendererID` and `GetDepthAttachmentRendererID` to use `void*` for consistent API support.
- Simplified Vulkan frame synchronization by consolidating semaphore logic.
- Restructured `VulkanRenderer` frame management logic for better modularity and reusability.
- Enhanced `EditorLayer` update logic to streamline rendering pipeline steps.

+++ src/renderer/OpenGLRenderer.cpp
@@ -335,7 +335,7 @@ void OpenGLRenderer::RenderScene(Scene* scene, int renderMode) {
             // --- INJECTION DU CSM (CASCADED SHADOW MAPS) ---
             glActiveTexture(GL_TEXTURE15);
             // ATTENTION : C'est un GL_TEXTURE_2D_ARRAY maintenant !
-            glBindTexture(GL_TEXTURE_2D_ARRAY, m_Data->ShadowFramebuffer->GetDepthAttachmentRendererID());
+            glBindTexture(GL_TEXTURE_2D_ARRAY, (GLuint)(uintptr_t)m_Data->ShadowFramebuffer->GetDepthAttachmentRendererID());
             activeShader->SetInt("uShadowMap", 15);
 
             // ==========================================================
